requirements.txt
wheel
tensorflow-cpu
Flask
python-dotenv
Pillow
gunicorn